nsexec.c: Build veth names using libuuid
Before this change, veth names were build using pid in the host ns, and
eth0 in the new netns. But, always using eth0 may clash with the other
interfaces when nsexec is executed in parallel.

Now, build a new uuid
using libuuid, and grab the first 8 characters from the new uuid,
splitting the first 4 of them for host veth, and the other 4 to use as
veth name of the newly created netns veth peer.

Signed-off-by: Marcos Paulo de Souza <marcos.souza.org@gmail.com>